nvme: actually pass parameters to nvme_admin_set_feature()
align len parameter type
noticed by waddlesplash, thanks!

Reviewed-on: https://review.haiku-os.org/c/haiku/+/9199

+++ src/add-ons/kernel/drivers/disk/nvme/nvme_disk.cpp
@@ -286,7 +286,7 @@ nvme_disk_init_device(void* _info, void** _cookie)
 					psd.apw);
 			}
 
-			size_t tableSize = 32 * sizeof(uint64);
+			uint32_t tableSize = 32 * sizeof(uint64);
 			uint64* table = (uint64*)malloc(tableSize);
 			memset(table, 0, tableSize);
